changed store architecture
